requirements.txt
-e git+https://github.com/ericsomdahl/python-bittrex.git@0.2.0#egg=python-bittrex
SQLAlchemy==1.1.14
python-telegram-bot==8.1.1
arrow==0.10.0
requests==2.18.4
urllib3==1.22
wrapt==1.10.11
pandas==0.20.3
scikit-learn==0.19.0
scipy==0.19.1
jsonschema==2.6.0
numpy==1.13.3
TA-Lib==0.4.10
pytest==3.2.3
pytest-mock==1.6.3
pytest-cov==2.5.1
hyperopt==0.1
# do not upgrade networkx before this is fixed https://github.com/hyperopt/hyperopt/issues/325
networkx==1.11
tabulate==0.8.1
pycrypto
progressbar2==3.34.3
tqdm==4.19.4

# Required for plotting data
#matplotlib==2.1.0
#PYQT5==5.9
